Split routes out from the API app 🎨
no issue

- Use the same pattern as the blog app
- General cleanup/unification across all of the `app.js` files
